+++ packages/components/src/integration/iot-status-grid/iot-status-grid.spec.component.ts
@@ -2,6 +2,7 @@ import { renderChart } from '../../testing/renderChart';
 import { mockLatestValueResponse } from '../../testing/mocks/mockGetAggregatedOrRawResponse';
 import { mockGetAssetSummary } from '../../testing/mocks/mockGetAssetSummaries';
 import { COMPARISON_OPERATOR } from '@synchro-charts/core';
+import { mockGetAssetModelSummary } from '../../testing/mocks/mockGetAssetModelSummary';
 
 const SECOND_IN_MS = 1000;
 
@@ -16,11 +17,15 @@ describe('status grid', () => {
   before(() => {
     cy.intercept('/properties/latest?*', (req) => {
       req.reply(mockLatestValueResponse());
-    });
+    }).as('getAggregates');
 
     cy.intercept(`/assets/${assetId}`, (req) => {
-      req.reply(mockGetAssetSummary({ assetModelId, id: assetId }));
-    });
+      req.reply(mockGetAssetSummary({ assetModelId, assetId }));
+    }).as('getAssetSummary');
+
+    cy.intercept(`/asset-models/${assetModelId}`, (req) => {
+      req.reply(mockGetAssetModelSummary({ assetModelId }));
+    }).as('getAssetModels');
   });
 
   it('renders', () => {
@@ -31,8 +36,34 @@ describe('status grid', () => {
       annotations: { y: [{ color: '#FF0000', comparisonOperator: COMPARISON_OPERATOR.GREATER_THAN, value: 25 }] },
     });
 
-    cy.wait(SECOND_IN_MS * 2);
+    cy.wait(['@getAggregates', '@getAssetSummary', '@getAssetModels']);
 
     cy.matchImageSnapshot(snapshotOptions);
+  });
+
+  it('renders passes all props to synchro-charts', () => {
+    const props = {
+      widgetId: '123',
+      viewport: { duration: '1m' },
+      annotations: { y: [{ color: '#FF0000', comparisonOperator: COMPARISON_OPERATOR.GREATER_THAN, value: 25 }] },
+      isEditing: false,
+    };
+
+    renderChart({
+      chartType: 'iot-status-grid',
+      settings: { resolution: '0' },
+      ...props,
+    });
+
+    cy.wait(SECOND_IN_MS * 2);
+
+    cy.get('sc-status-grid').should((e) => {
+      const [chart] = e.get();
+      (Object.keys(props) as Array<keyof typeof props>).forEach((prop) => {
+        const value = chart[prop as keyof HTMLScStatusGridElement];
+        const passedInValue = props[prop];
+        expect(value).to.deep.equal(passedInValue);
+      });
+    });
   });
 });