+++ packages/components/src/integration/iot-status-timeline/iot-status-timeline.spec.component.ts
@@ -2,6 +2,7 @@ import { renderChart } from '../../testing/renderChart';
 import { mockGetAggregatedOrRawResponse } from '../../testing/mocks/mockGetAggregatedOrRawResponse';
 import { mockGetAssetSummary } from '../../testing/mocks/mockGetAssetSummaries';
 import { COMPARISON_OPERATOR } from '@synchro-charts/core';
+import { mockGetAssetModelSummary } from '../../testing/mocks/mockGetAssetModelSummary';
 
 const SECOND_IN_MS = 1000;
 
@@ -32,11 +33,15 @@ describe('status timeline', () => {
           })
         );
       }
-    });
+    }).as('getAggregates');
 
     cy.intercept(`/assets/${assetId}`, (req) => {
-      req.reply(mockGetAssetSummary({ assetModelId, id: assetId }));
-    });
+      req.reply(mockGetAssetSummary({ assetModelId, assetId }));
+    }).as('getAssetSummary');
+
+    cy.intercept(`/asset-models/${assetModelId}`, (req) => {
+      req.reply(mockGetAssetModelSummary({ assetModelId }));
+    }).as('getAssetModels');
   });
 
   it('renders', () => {
@@ -46,8 +51,34 @@ describe('status timeline', () => {
       annotations: { y: [{ color: '#FF0000', comparisonOperator: COMPARISON_OPERATOR.GREATER_THAN, value: 26 }] },
     });
 
-    cy.wait(SECOND_IN_MS * 2);
+    cy.wait(['@getAggregates', '@getAssetSummary', '@getAssetModels']);
 
     cy.matchImageSnapshot(snapshotOptions);
+  });
+
+  it('renders passes all props to synchro-charts', () => {
+    const props = {
+      widgetId: '123',
+      viewport: { duration: '1m' },
+      annotations: { y: [{ color: '#FF0000', comparisonOperator: COMPARISON_OPERATOR.GREATER_THAN, value: 26 }] },
+      isEditing: false,
+    };
+
+    renderChart({
+      chartType: 'iot-status-timeline',
+      settings: { resolution: '0' },
+      ...props,
+    });
+
+    cy.wait(SECOND_IN_MS * 2);
+
+    cy.get('sc-status-timeline').should((e) => {
+      const [chart] = e.get();
+      (Object.keys(props) as Array<keyof typeof props>).forEach((prop) => {
+        const value = chart[prop as keyof HTMLScStatusTimelineElement];
+        const passedInValue = props[prop];
+        expect(value).to.deep.equal(passedInValue);
+      });
+    });
   });
 });